+++ pkg/ingester/user_metrics_metadata_test.go
@@ -20,6 +20,8 @@ func TestUserMetricsMetadata(t *testing.T) {
 		errContains string
 	}
 
+	errorSamplers := newIngesterErrSamplers(0)
+
 	tests := map[string]struct {
 		maxMetadataPerUser   int
 		maxMetadataPerMetric int
@@ -92,7 +94,7 @@ func TestUserMetricsMetadata(t *testing.T) {
 				nil,
 			)
 
-			mm := newMetadataMap(limiter, metrics, "test")
+			mm := newMetadataMap(limiter, metrics, errorSamplers, "test")
 
 			// Attempt to add all metadata
 			for _, i := range testData.inputMetadata {